test(adapter-hermes): split Hermes adapter test suite
