+++ packages/adapter-hermes/test/hermes-adapter.part-02.test.ts
@@ -1,0 +1,79 @@
+import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
+import { existsSync, mkdirSync, mkdtempSync, readdirSync, readFileSync, rmSync, writeFileSync } from 'node:fs';
+import { join, resolve } from 'node:path';
+import { tmpdir } from 'node:os';
+import { spawnSync } from 'node:child_process';
+vi.mock('@origintrail-official/dkg-core', async () => {
+  const actual = await vi.importActual<typeof import('@origintrail-official/dkg-core')>(
+    '@origintrail-official/dkg-core',
+  );
+  return {
+    ...actual,
+    resolveDkgConfigHome: vi.fn((opts) => actual.resolveDkgConfigHome(opts)),
+    resolveDkgHome: vi.fn((opts) => actual.resolveDkgHome(opts)),
+  };
+});
+import { resolveDkgHome } from '@origintrail-official/dkg-core';
+import { HermesAdapterPlugin } from '../src/HermesAdapterPlugin.js';
+import { registerHermesRoutes } from '../src/hermes-routes.js';
+import { HermesDkgClient, redact } from '../src/dkg-client.js';
+import {
+  disconnectHermesProfile,
+  planHermesSetup,
+  runDoctor,
+  runDisconnect,
+  runReconnect,
+  resolveHermesProfile,
+  runSetup,
+  runUninstall,
+  runVerify,
+  setupHermesProfile,
+  uninstallHermesProfile,
+  verifyHermesProfile,
+} from '../src/setup.js';
+
+afterEach(() => {
+  vi.unstubAllGlobals();
+  vi.restoreAllMocks();
+});
+import { createTrackingApi, trackingRes, type TrackingApi } from './hermes-adapter.shared';
+
+
+
+describe('HermesAdapterPlugin', () => {
+
+
+  it('registers HTTP routes on first call', () => {
+    const plugin = new HermesAdapterPlugin();
+    const api = createTrackingApi();
+
+    plugin.register(api);
+
+    expect(api.registerHttpRouteCalls).toHaveLength(1);
+    expect([...api.routes.keys()].sort()).toEqual([
+      'GET /api/hermes/status',
+    ]);
+  });
+
+  it('registers session_end lifecycle hook', () => {
+    const plugin = new HermesAdapterPlugin();
+    const api = createTrackingApi();
+
+    plugin.register(api);
+
+    expect(api.registerHookCalls.some(
+      (c: any) => c.event === 'session_end' && c.meta?.name === 'hermes-adapter-stop',
+    )).toBe(true);
+  });
+
+  it('skips route registration on second call (idempotent)', () => {
+    const plugin = new HermesAdapterPlugin();
+    const api = createTrackingApi();
+
+    plugin.register(api);
+    plugin.register(api);
+
+    expect(api.registerHttpRouteCalls).toHaveLength(1);
+  });
+
+});
