+++ packages/adapter-hermes/test/hermes-adapter.part-03.test.ts
@@ -1,0 +1,61 @@
+import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
+import { existsSync, mkdirSync, mkdtempSync, readdirSync, readFileSync, rmSync, writeFileSync } from 'node:fs';
+import { join, resolve } from 'node:path';
+import { tmpdir } from 'node:os';
+import { spawnSync } from 'node:child_process';
+vi.mock('@origintrail-official/dkg-core', async () => {
+  const actual = await vi.importActual<typeof import('@origintrail-official/dkg-core')>(
+    '@origintrail-official/dkg-core',
+  );
+  return {
+    ...actual,
+    resolveDkgConfigHome: vi.fn((opts) => actual.resolveDkgConfigHome(opts)),
+    resolveDkgHome: vi.fn((opts) => actual.resolveDkgHome(opts)),
+  };
+});
+import { resolveDkgHome } from '@origintrail-official/dkg-core';
+import { HermesAdapterPlugin } from '../src/HermesAdapterPlugin.js';
+import { registerHermesRoutes } from '../src/hermes-routes.js';
+import { HermesDkgClient, redact } from '../src/dkg-client.js';
+import {
+  disconnectHermesProfile,
+  planHermesSetup,
+  runDoctor,
+  runDisconnect,
+  runReconnect,
+  resolveHermesProfile,
+  runSetup,
+  runUninstall,
+  runVerify,
+  setupHermesProfile,
+  uninstallHermesProfile,
+  verifyHermesProfile,
+} from '../src/setup.js';
+
+afterEach(() => {
+  vi.unstubAllGlobals();
+  vi.restoreAllMocks();
+});
+import { createTrackingApi, trackingRes, type TrackingApi } from './hermes-adapter.shared';
+
+
+
+describe('GET /api/hermes/status', () => {
+
+
+  it('returns adapter status JSON', async () => {
+    const api = createTrackingApi();
+    registerHermesRoutes(api);
+    const handler = api.routes.get('GET /api/hermes/status')!;
+    const { res, calls } = trackingRes();
+
+    await handler({}, res);
+
+    expect(calls.some(c =>
+      c.json?.adapter === 'hermes' &&
+      c.json?.framework === 'hermes-agent' &&
+      c.json?.status === 'connected',
+    )).toBe(true);
+  });
+
+});
