+++ packages/adapter-hermes/test/hermes-adapter.shared.ts
@@ -1,0 +1,69 @@
+import type { DaemonPluginApi } from '../src/types.js';
+
+interface TrackingApi extends DaemonPluginApi {
+  routes: Map<string, (req: any, res: any) => Promise<void>>;
+  hooks: Map<string, () => Promise<void>>;
+  registerHttpRouteCalls: any[];
+  registerHookCalls: any[];
+}
+
+function createTrackingApi(): TrackingApi {
+  const routes = new Map<string, (req: any, res: any) => Promise<void>>();
+  const hooks = new Map<string, () => Promise<void>>();
+  const registerHttpRouteCalls: any[] = [];
+  const registerHookCalls: any[] = [];
+
+  const storeChatTurnCalls: any[][] = [];
+  const importMemoriesCalls: any[][] = [];
+  let storeChatTurnError: Error | null = null;
+  let importMemoriesError: Error | null = null;
+
+  return {
+    routes,
+    hooks,
+    registerHttpRouteCalls,
+    registerHookCalls,
+    registerHttpRoute: (opts: any) => {
+      registerHttpRouteCalls.push(opts);
+      routes.set(`${opts.method} ${opts.path}`, opts.handler);
+    },
+    registerHook: (event: string, handler: any, meta?: any) => {
+      registerHookCalls.push({ event, handler, meta });
+      hooks.set(event, handler);
+    },
+    logger: {
+      info: () => {},
+      warn: () => {},
+      debug: () => {},
+    },
+    agent: {
+      query: async () => {},
+      share: async () => {},
+      importMemories: async (...args: any[]) => {
+        importMemoriesCalls.push(args);
+        if (importMemoriesError) throw importMemoriesError;
+        return undefined;
+      },
+      storeChatTurn: async (...args: any[]) => {
+        storeChatTurnCalls.push(args);
+        if (storeChatTurnError) throw storeChatTurnError;
+        return undefined;
+      },
+      _storeChatTurnCalls: storeChatTurnCalls,
+      _importMemoriesCalls: importMemoriesCalls,
+      _setStoreChatTurnError: (e: Error | null) => { storeChatTurnError = e; },
+      _setImportMemoriesError: (e: Error | null) => { importMemoriesError = e; },
+    },
+  } as any;
+}
+
+function trackingRes() {
+  const calls: { status?: number; json?: any }[] = [];
+  const res: any = {};
+  res.status = (code: number) => { calls.push({ status: code }); return res; };
+  res.json = (body: any) => { calls.push({ json: body }); return res; };
+  return { res, calls };
+}
+
+export { createTrackingApi, trackingRes };
+export type { TrackingApi };
